egyptian-tts-backend/requirements.txt
arabic-reshaper==3.0.0
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==36.0.2
csvw==3.5.1
decorator==5.2.1
dlinfo==2.0.0
Flask==3.1.0
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
language-tags==1.2.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
msgpack==1.1.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
phonemizer==3.3.0
platformdirs==4.3.8
pooch==1.8.2
PyArabic==0.6.15
pycparser==2.22
pydub==0.25.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.3
segments==2.3.0
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.40
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
Werkzeug==3.1.3
